requirements.txt
python-telegram-bot==22.6
python-dotenv==1.2.2
httpx==0.28.1
APScheduler==3.11.2
pytz==2026.1.post1
tzlocal==5.3.1
flask==3.1.0